refactor: split `log_wrapper_helpers.h`
I will need some of these functions in libraries that do not depend on
`google_cloud_cpp_grpc_utils`.  I also improved the tests by a tiny
amount.
