+++ google/cloud/google_cloud_cpp_common.bzl
@@ -40,6 +40,8 @@ google_cloud_cpp_common_hdrs = [
     "internal/compiler_info.h",
     "internal/compute_engine_util.h",
     "internal/credentials_impl.h",
+    "internal/debug_future_status.h",
+    "internal/debug_string.h",
     "internal/diagnostics_pop.inc",
     "internal/diagnostics_push.inc",
     "internal/disable_deprecation_warnings.inc",
@@ -102,6 +104,8 @@ google_cloud_cpp_common_srcs = [
     "internal/compiler_info.cc",
     "internal/compute_engine_util.cc",
     "internal/credentials_impl.cc",
+    "internal/debug_future_status.cc",
+    "internal/debug_string.cc",
     "internal/error_context.cc",
     "internal/filesystem.cc",
     "internal/format_time_point.cc",
