+++ google/cloud/google_cloud_cpp_common_unit_tests.bzl
@@ -32,6 +32,8 @@ google_cloud_cpp_common_unit_tests = [
     "internal/compiler_info_test.cc",
     "internal/compute_engine_util_test.cc",
     "internal/credentials_impl_test.cc",
+    "internal/debug_future_status_test.cc",
+    "internal/debug_string_test.cc",
     "internal/error_context_test.cc",
     "internal/filesystem_test.cc",
     "internal/format_time_point_test.cc",
