+++ google/cloud/google_cloud_cpp_grpc_utils.bzl
@@ -49,6 +49,8 @@ google_cloud_cpp_grpc_utils_hdrs = [
     "internal/async_streaming_write_rpc_logging.h",
     "internal/background_threads_impl.h",
     "internal/completion_queue_impl.h",
+    "internal/debug_string_protobuf.h",
+    "internal/debug_string_status.h",
     "internal/default_completion_queue_impl.h",
     "internal/extract_long_running_result.h",
     "internal/grpc_access_token_authentication.h",
@@ -58,7 +60,6 @@ google_cloud_cpp_grpc_utils_hdrs = [
     "internal/grpc_request_metadata.h",
     "internal/grpc_service_account_authentication.h",
     "internal/log_wrapper.h",
-    "internal/log_wrapper_helpers.h",
     "internal/minimal_iam_credentials_stub.h",
     "internal/populate_grpc_options.h",
     "internal/resumable_streaming_read_rpc.h",
@@ -82,6 +83,8 @@ google_cloud_cpp_grpc_utils_srcs = [
     "internal/async_connection_ready.cc",
     "internal/async_polling_loop.cc",
     "internal/background_threads_impl.cc",
+    "internal/debug_string_protobuf.cc",
+    "internal/debug_string_status.cc",
     "internal/default_completion_queue_impl.cc",
     "internal/extract_long_running_result.cc",
     "internal/grpc_access_token_authentication.cc",
@@ -90,7 +93,6 @@ google_cloud_cpp_grpc_utils_srcs = [
     "internal/grpc_impersonate_service_account.cc",
     "internal/grpc_request_metadata.cc",
     "internal/grpc_service_account_authentication.cc",
-    "internal/log_wrapper_helpers.cc",
     "internal/minimal_iam_credentials_stub.cc",
     "internal/populate_grpc_options.cc",
     "internal/streaming_read_rpc.cc",
