+++ google/cloud/google_cloud_cpp_grpc_utils_unit_tests.bzl
@@ -36,13 +36,14 @@ google_cloud_cpp_grpc_utils_unit_tests = [
     "internal/async_streaming_write_rpc_impl_test.cc",
     "internal/async_streaming_write_rpc_logging_test.cc",
     "internal/background_threads_impl_test.cc",
+    "internal/debug_string_protobuf_test.cc",
+    "internal/debug_string_status_test.cc",
     "internal/extract_long_running_result_test.cc",
     "internal/grpc_access_token_authentication_test.cc",
     "internal/grpc_async_access_token_cache_test.cc",
     "internal/grpc_channel_credentials_authentication_test.cc",
     "internal/grpc_request_metadata_test.cc",
     "internal/grpc_service_account_authentication_test.cc",
-    "internal/log_wrapper_helpers_test.cc",
     "internal/log_wrapper_test.cc",
     "internal/minimal_iam_credentials_stub_test.cc",
     "internal/populate_grpc_options_test.cc",
