+++ google/cloud/internal/debug_future_status.h
@@ -12,13 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_LOG_WRAPPER_HELPERS_H
-#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_LOG_WRAPPER_HELPERS_H
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_DEBUG_FUTURE_STATUS_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_DEBUG_FUTURE_STATUS_H
 
-#include "google/cloud/status.h"
 #include "google/cloud/tracing_options.h"
 #include "google/cloud/version.h"
-#include <google/protobuf/message.h>
 #include <future>
 #include <string>
 
@@ -27,22 +25,11 @@ namespace cloud {
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
 namespace internal {
 
-std::string DebugString(google::protobuf::Message const& m,
-                        TracingOptions const& options);
-
-std::string DebugString(Status const& status, TracingOptions const& options);
-
-std::string DebugString(std::string s, TracingOptions const& options);
-
 char const* DebugFutureStatus(std::future_status s);
-
-// Create a unique ID that can be used to match asynchronous requests/response
-// pairs.
-std::string RequestIdForLogging();
 
 }  // namespace internal
 GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
 }  // namespace cloud
 }  // namespace google
 
-#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_LOG_WRAPPER_HELPERS_H
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_INTERNAL_DEBUG_FUTURE_STATUS_H
