+++ google/cloud/internal/debug_future_status_test.cc
@@ -1,0 +1,42 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#include "google/cloud/internal/debug_future_status.h"
+#include <gmock/gmock.h>
+
+namespace google {
+namespace cloud {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+namespace internal {
+namespace {
+
+TEST(DebugString, FutureStatus) {
+  struct Case {
+    std::future_status actual;
+    std::string expected;
+  } cases[] = {
+      {std::future_status::deferred, "deferred"},
+      {std::future_status::timeout, "timeout"},
+      {std::future_status::ready, "ready"},
+  };
+  for (auto const& c : cases) {
+    EXPECT_EQ(c.expected, DebugFutureStatus(c.actual));
+  }
+}
+
+}  // namespace
+}  // namespace internal
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
+}  // namespace cloud
+}  // namespace google
