+++ google/cloud/internal/log_wrapper.h
@@ -18,8 +18,11 @@
 #include "google/cloud/completion_queue.h"
 #include "google/cloud/future.h"
 #include "google/cloud/grpc_error_delegate.h"
+#include "google/cloud/internal/debug_future_status.h"
+#include "google/cloud/internal/debug_string.h"
+#include "google/cloud/internal/debug_string_protobuf.h"
+#include "google/cloud/internal/debug_string_status.h"
 #include "google/cloud/internal/invoke_result.h"
-#include "google/cloud/internal/log_wrapper_helpers.h"
 #include "google/cloud/log.h"
 #include "google/cloud/status_or.h"
 #include "google/cloud/tracing_options.h"
